# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Aster DNS API
  version: 1.0.0
  description: |
    REST API for the self-hosted Aster DNS control plane. DNS Providers are the
    authoritative source for zones and records. Cookie-authenticated mutations
    require both a trusted Origin and the X-CSRF-Token header.
servers:
  - url: /
...
paths:
...
  /api/v1/provider-types:
    get:
      tags: [Providers]
      summary: List registered Provider capabilities and input schemas
      operationId: listProviderTypes
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Provider type catalog
          content:
            application/json:
              schema:
                type: object
                required: [provider_types]
                properties:
                  provider_types:
                    type: array
                    items: { $ref: '#/components/schemas/ProviderType' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/provider-accounts:
    get:
      tags: [Providers]
      summary: List safe Provider account state
      operationId: listProviderAccounts
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Provider accounts without credential material
          content:
            application/json:
              schema:
                type: object
                required: [provider_accounts]
                properties:
                  provider_accounts:
                    type: array
                    items: { $ref: '#/components/schemas/ProviderAccount' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Providers]
      summary: Create a Provider account
      description: Admin only. Credentials are write-only.
      operationId: createProviderAccount
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProviderAccount' }
      responses:
        '201': { $ref: '#/components/responses/ProviderAccount' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/provider-accounts/{id}:
    get:
      tags: [Providers]
      summary: Read safe Provider account state
      operationId: getProviderAccount
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      responses:
        '200': { $ref: '#/components/responses/ProviderAccount' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      tags: [Providers]
      summary: Update Provider account metadata/options
      description: Admin only. Credentials cannot be changed here.
      operationId: updateProviderAccount
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string }
                description: { type: string }
                enabled: { type: boolean }
                options: { type: object, additionalProperties: true }
      responses:
        '200': { $ref: '#/components/responses/ProviderAccount' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Providers]
      summary: Delete a Provider account
      description: Admin only. Does not delete the Provider zone.
      operationId: deleteProviderAccount
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Deleted }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/provider-accounts/{id}/credentials:
    post:
      tags: [Providers]
      summary: Replace Provider credentials
      description: Admin only. Existing secret values are never readable.
      operationId: replaceProviderCredentials
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [credentials]
              properties:
                credentials:
                  type: object
                  writeOnly: true
                  additionalProperties: true
      responses:
        '200': { $ref: '#/components/responses/ProviderAccount' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/provider-accounts/{id}/validate:
    post:
      tags: [Providers]
      summary: Validate an account using a minimal read-only Provider request
      operationId: validateProviderAccount
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200': { $ref: '#/components/responses/ProviderAccount' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/provider-accounts/{id}/sync-zones:
    post:
      tags: [Providers]
      summary: Synchronize the account Zone index
      operationId: syncProviderZones
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200':
          description: Synchronization result
          content:
            application/json:
              schema:
                type: object
                required: [status, zone_count]
                properties:
                  status: { const: synchronized }
                  zone_count: { type: integer, minimum: 0 }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-aster_session
      description: Opaque session cookie. HTTP deployments use the non-Host-prefixed equivalent.
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: Must match the readable CSRF cookie for the current server-side session.
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
    EmptyJSON:
      required: false
      content:
        application/json:
          schema: { type: object, maxProperties: 0 }
  responses:
    Error:
      description: Stable safe error. Details never contain Provider secrets or raw upstream payloads.
      headers:
        X-Request-ID:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ProviderAccount:
      description: Safe Provider account state without credential material
      content:
        application/json:
          schema:
            type: object
            required: [provider_account]
            properties:
              provider_account: { $ref: '#/components/schemas/ProviderAccount' }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ItemError' }
    ItemError:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
...
    DescriptorOption:
      type: object
      required: [value, label]
      properties:
        value: { type: string }
        label: { type: string }
    DescriptorCondition:
      type: object
      required: [field, values]
      properties:
        field: { type: string }
        values:
          type: array
          items: { type: string }
    DescriptorField:
      type: object
      required: [key, label, type, secret, required]
      properties:
        key: { type: string }
        label: { type: string }
        type: { type: string, enum: [string, integer, boolean, enum, string_list] }
        secret: { type: boolean }
        required: { type: boolean }
        placeholder: { type: string }
        description: { type: string }
        options:
          type: array
          items: { $ref: '#/components/schemas/DescriptorOption' }
        minimum: { type: integer }
        maximum: { type: integer }
    ExtensionFieldDescriptor:
      type: object
      required: [namespace, scope, key, label, type, read_only, required]
      properties:
        namespace: { type: string }
        scope: { type: string, enum: [zone, record_set, record_entry] }
        key: { type: string }
        label: { type: string }
        type: { type: string, enum: [string, integer, boolean, enum, string_list] }
        read_only: { type: boolean }
        required: { type: boolean }
        applicable_when:
          type: array
          items: { $ref: '#/components/schemas/DescriptorCondition' }
        required_when:
          type: array
          items: { $ref: '#/components/schemas/DescriptorCondition' }
        options:
          type: array
          items: { $ref: '#/components/schemas/DescriptorOption' }
        minimum: { type: integer }
        maximum: { type: integer }
    ProviderCapabilities:
      type: object
      required: [supported_record_types, native_record_granularity, supports_proxy, supports_routing_line, supports_weight, supports_record_status, supports_dnssec, supports_native_batch, supports_comments, extension_fields]
      properties:
        supported_record_types:
          type: array
          items: { $ref: '#/components/schemas/RecordType' }
        min_ttl: { type: integer, minimum: 0 }
        max_ttl: { type: integer, minimum: 1 }
        native_record_granularity: { type: string, enum: [record_set, record_entry] }
        supports_proxy: { type: boolean }
        supports_routing_line: { type: boolean }
        supports_weight: { type: boolean }
        supports_record_status: { type: boolean }
        supports_dnssec: { type: boolean }
        supports_native_batch: { type: boolean }
        supports_comments: { type: boolean }
        extension_fields:
          type: array
          items: { $ref: '#/components/schemas/ExtensionFieldDescriptor' }
    ProviderType:
      type: object
      required: [type, display_name, display_names, credential_fields, account_options, capabilities]
      properties:
        type: { type: string, enum: [huawei, aliyun, tencent, cloudflare] }
        display_name: { type: string }
        display_names:
          type: object
          minProperties: 1
          additionalProperties: { type: string }
          description: Localized display names keyed by BCP 47 language tag.
        documentation_url: { type: string, format: uri }
        credential_fields:
          type: array
          items: { $ref: '#/components/schemas/DescriptorField' }
        account_options:
          type: array
          items: { $ref: '#/components/schemas/DescriptorField' }
        capabilities: { $ref: '#/components/schemas/ProviderCapabilities' }
    ValidationStatus:
      type: string
      enum: [unconfigured, pending, valid, invalid, error]
    ProviderAccount:
      type: object
      description: Never contains plaintext credentials, ciphertext, nonce, or encryption key metadata.
      required: [id, provider_type, name, description, enabled, options, credential_configured, credential_revision, validation_status, zone_count, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        provider_type: { type: string }
        name: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        options: { type: object, additionalProperties: true }
        credential_configured: { type: boolean }
        credential_revision: { type: integer, minimum: 0 }
        validation_status: { $ref: '#/components/schemas/ValidationStatus' }
        last_validated_at: { type: string, format: date-time }
        last_validation_error_code: { type: string }
        last_zone_sync_at: { type: string, format: date-time }
        zone_count: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateProviderAccount:
      type: object
      additionalProperties: false
      required: [provider_type, name]
      properties:
        provider_type: { type: string }
        name: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        options: { type: object, additionalProperties: true }
        credentials: { type: object, writeOnly: true, additionalProperties: true }
...
    RecordType:
      type: string
      enum: [A, AAAA, CNAME, TXT, MX, NS, SRV, CAA]
...